build: update vscode config to use golangci-lint as lint tool
Rclone upstream runs golangci-lint as linter in a GitHub workflow, as a pull request
check, configured with a .golangci.yml file.

VSCode by default uses staticcheck as a lint tool, in addition to the gopls language
server which also is also used for some code analysis (in addition to intellisense and
code navigation). This commit changes VSCode to use golangci-lint instead of staticcheck,
based on the same .golangci.yml configuration file. This helps catching all linter issues
locally, before creating a pull request upstream. Note that VSCode may report some
additional issues, not coming from golangci-lint but from gopls, and these may or may not
describe changes that are wanted upstream. Explicit running a linting task from a clean
workspace makes sure only golangci-lint issues are reported.

This also changes the default EOL format for new files from native OS to forced LF.
Lint checks from gofmt and goimports (which in reality is a wrapper of gofmt)
which are ran by golangci-lint, assumes sources files always use LF line endings and
on Windows files with CRLF will be reported with error:
"File is not `goimports`-ed (goimports)". To avoid this, the repository should be checked
out with its original EOL, assuming it is normalized to LF. Configuring VSCode to create
any new files with LF, is therefore also important. To be on the safe side, to avoid
any files created with, or converted to, CRLF locally being pushed, git should also
be configured to convert any CRLF to LF on push, by using the autocrlf input strategy.
So run git checkout and configuration like this:

    git clone git@github.com:user/repo.git --config core.autocrlf=input
    git config core.autocrlf input